Adding CC0 license, repository field
Hi, I'd like to use this, but one of the requirements for my job is that dependencies need to have a "compatible" license. MIT fits the bill and is also [incredibly short and simple](http://opensource.org/licenses/MIT), so that's what I'd recommend. If you'd prefer a different one, though, go for it. I'd just like to see something listed ;)
(Update: switched license to CC0)

Oh, and I'm adding a repository field to make it easy to get to the source from npm.

switched license to CC0